warpdrive platform-specific constructors
