+++ lib/warpdrived/desktop.go
@@ -1,0 +1,5 @@
+package warpdrived
+
+func Desktop() *Server {
+	return &Server{}
+}
